pyproject.toml
[tool.poetry]
name = "arbitrage-bot"
version = "0.1.0"
description = "Solana DEX arbitrage bot scaffold"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
solana = "^0.33.0"
httpx = "^0.27.0"
python-igraph = "^0.11.4"
numpy = "^1.26.0"
scipy = "^1.13.0"
scikit-learn = "^1.5.0"
APScheduler = "^3.10.4"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"